final/server/requirements.txt
fastmcp
requests
python-dotenv
mcp
google-cloud-bigquery

